Changes to code and test for breaking out transformers.
