+++ healthcareai/common/dataframe_transformers/data_frame_drop_nan.py
@@ -1,0 +1,21 @@
+import numpy as np
+import pandas as pd
+
+from sklearn.base import TransformerMixin
+from imblearn.over_sampling import RandomOverSampler
+from imblearn.under_sampling import RandomUnderSampler
+from sklearn.preprocessing import StandardScaler
+
+class DataFrameDropNaN(TransformerMixin):
+    """Remove NaN values. Columns that are NaN or None are removed."""
+
+    def __init__(self):
+        pass
+
+    def fit(self, X, y=None):
+        return self
+
+    def transform(self, X, y=None):
+        # Uses pandas.DataFrame.dropna function where axis=1 is column action, and
+        # how='all' requires all the values to be NaN or None to be removed.
+        return X.dropna(axis=1, how='all')
